+++ samples/system-test/files.test.js
@@ -1,10 +1,11 @@
 /**
- * Copyright 2017, Google, Inc.
+ * Copyright 2019 Google LLC
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -36,6 +37,7 @@ const copiedFileName = 'test3.txt';
 const signedFileName = 'signed-upload.txt';
 const kmsKeyName = process.env.GOOGLE_CLOUD_KMS_KEY_US;
 const filePath = path.join(cwd, 'resources', fileName);
+const folderPath = path.join(cwd, 'resources');
 const downloadFilePath = path.join(cwd, 'downloaded.txt');
 const cmd = `node files.js`;
 
@@ -54,7 +56,7 @@ after(async () => {
 });
 
 it('should upload a file', async () => {
-  const output = execSync(`${cmd} upload ${bucketName} ${filePath}`);
+  const output = execSync(`node uploadFile.js ${bucketName} ${filePath}`);
   assert.match(output, new RegExp(`${filePath} uploaded to ${bucketName}.`));
   const [exists] = await bucket.file(fileName).exists();
   assert.strictEqual(exists, true);
@@ -70,6 +72,49 @@ it('should upload a file with a kms key', async () => {
   );
   const [exists] = await bucket.file(fileName).exists();
   assert.strictEqual(exists, true);
+});
+
+it('should upload a local directory', done => {
+  const output = execSync(
+    `node uploadDirectory.js ${bucketName} ${folderPath}`
+  );
+
+  const fileList = [];
+  getFileList(folderPath);
+
+  function getFileList(directory) {
+    const items = fs.readdirSync(directory);
+    items.forEach(item => {
+      const fullPath = path.join(directory, item);
+      const stat = fs.lstatSync(fullPath);
+      if (stat.isFile()) {
+        fileList.push(fullPath);
+      } else {
+        getFileList(fullPath);
+      }
+    });
+  }
+
+  assert.match(
+    output,
+    new RegExp(
+      `${fileList.length} files uploaded to ${bucketName} successfully.`
+    )
+  );
+
+  Promise.all(
+    fileList.map(file =>
+      bucket
+        .file(path.relative(path.dirname(folderPath), file).replace(/\\/g, '/'))
+        .exists()
+    )
+  ).then(resps => {
+    const ctr = resps.reduce((acc, cur) => {
+      return acc + cur[0];
+    }, 0);
+    assert.strictEqual(ctr, fileList.length);
+    done();
+  }, assert.ifError);
 });
 
 it('should download a file', () => {
@@ -203,6 +248,23 @@ it('should get metadata for a file', () => {
   );
   assert.match(output, new RegExp(`File: ${copiedFileName}`));
   assert.match(output, new RegExp(`Bucket: ${bucketName}`));
+});
+
+it('should set metadata for a file', () => {
+  // used in sample
+  const userMetadata = {
+    description: 'file description...',
+    modified: '1900-01-01',
+  };
+  const output = execSync(
+    `node fileSetMetadata.js ${bucketName} ${copiedFileName}`
+  );
+
+  assert.match(
+    output,
+    new RegExp(`description: '${userMetadata.description}'`)
+  );
+  assert.match(output, new RegExp(`modified: '${userMetadata.modified}'`));
 });
 
 it('should delete a file', async () => {